+++ src/features/apps/apiSlice.ts
@@ -125,6 +125,7 @@ export enum DocumentTypeText {
   CONFORMITY_DOCUMENT = 'ConformityDocument',
   DOCUMENTS = 'Documents',
   CONFORMITY_APPROVAL_BUSINESS_APPS = 'CONFORMITY_APPROVAL_BUSINESS_APPS',
+  CONFORMITY_APPROVAL_SERVICES = 'CONFORMITY_APPROVAL_SERVICES',
 }
 
 export type DocumentData = {
@@ -288,6 +289,7 @@ export const apiSlice = createApi({
             src: subscriptionItem.image
               ? `${getApiBase()}/api/apps/${
                   subscriptionItem.offerId
+                  // eslint-disable-next-line
                 }/appDocuments/${subscriptionItem.image}`
               : LogoGrayData,
           }