+++ CHANGELOG.md
@@ -40,6 +40,8 @@
   - fixed automatic upscrolling in form pages
 - Service Release Process
   - fixed automatic upscrolling in form pages
+- Service Management
+  - Added conformity document UI and functionality
 - Idp
   - Added snackbar for idp error